+++ jjava-jupyter/src/main/java/org/dflib/jjava/jupyter/kernel/magic/MagicTranspiler.java
@@ -15,8 +15,8 @@ public class MagicTranspiler {
     private static final Pattern UNESCAPED_QUOTE = Pattern.compile("(?<!\\\\)\"");
 
     // generated code templates
-    private static final String CELL_CALL_TEMPLATE = BaseNotebookStatics.class.getName() + ".cellMagic(%s,java.util.List.of(%s),%s);{};";
-    private static final String LINE_CALL_TEMPLATE = BaseNotebookStatics.class.getName() + ".lineMagic(%s,java.util.List.of(%s));{};";
+    private static final String CELL_CALL_TEMPLATE = BaseNotebookStatics.class.getName() + ".cellMagic(%s,java.util.List.of(%s),%s);";
+    private static final String LINE_CALL_TEMPLATE = BaseNotebookStatics.class.getName() + ".lineMagic(%s,java.util.List.of(%s));";
     private static final String DECODE_TEMPLATE = "new String(java.util.Base64.getDecoder().decode(\"%s\"))";
 
     public String transpileCell(ParsedCellMagic magic) {