+++ jjava-kernel/src/main/java/org/dflib/jjava/kernel/magics/ClasspathMagic.java
@@ -8,16 +8,16 @@ import java.io.File;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ClasspathMagic implements LineMagic<String, JavaKernel> {
+public class ClasspathMagic implements LineMagic<Void, JavaKernel> {
 
     @Override
-    public String eval(JavaKernel kernel, List<String> args) {
+    public Void eval(JavaKernel kernel, List<String> args) {
         String classpath = args.stream()
                 .flatMap(a -> PathsHandler.resolveGlobs(a).stream())
                 .map(p -> p.toAbsolutePath().toString())
                 .collect(Collectors.joining(File.pathSeparator));
 
         kernel.addToClasspath(classpath);
-        return classpath;
+        return null;
     }
 }